refactor: unixfs/cat.rs from the module
